Homeview ok, segue to base color ok

+++ Swatches/Model/Color.swift
@@ -11,9 +11,9 @@ struct Color {
     var colorName:String = ""
     var colorTag:String = ""
     var hexValue:String = ""
-    var red:Int = 0
-    var green:Int = 0
-    var blue:Int = 0
+    var red:Double = 0
+    var green:Double = 0
+    var blue:Double = 0
 }
